terraform: new output transform that isn't used yet
